put training task / regimen classes into their proper modules
